No faulty files on linux-aarch64

+++ recipe/build.sh
@@ -6,7 +6,9 @@ mkdir -p $PREFIX/share
 cp -R code-server $PREFIX/share/
 
 # Remove faulty files
-find $PREFIX/share -name node_gyp_bins | xargs rm -r
+if [[ "${target_platform}" != "linux-aarch64" ]]; then
+  find $PREFIX/share -name node_gyp_bins | xargs rm -r
+fi
 
 mkdir -p ${PREFIX}/bin
 mkdir -p ${PREFIX}/share/code-server/extensions
